+++ src/capnp/serialize.rs
@@ -16,12 +16,12 @@ use io;
 pub struct OwnedSpaceMessageReader {
     options : ReaderOptions,
     arena : Box<arena::ReaderArena>,
-    segment_slices : Vec<(uint, uint)>,
+    segment_slices : Vec<(usize, usize)>,
     owned_space : Vec<Word>,
 }
 
 impl MessageReader for OwnedSpaceMessageReader {
-    fn get_segment(&self, id : uint) -> &[Word] {
+    fn get_segment(&self, id : usize) -> &[Word] {
         let (a,b) = self.segment_slices.as_slice()[id];
         self.owned_space.slice(a, b)
     }
@@ -64,11 +64,11 @@ pub fn new_reader<U : std::io::Reader>(input_stream : &mut U,
         return invalid_input("too many segments");
     }
 
-    let mut more_sizes : Vec<u32> = Vec::with_capacity((segment_count & !1) as uint);
+    let mut more_sizes : Vec<u32> = Vec::with_capacity((segment_count & !1) as usize);
 
     if segment_count > 1 {
-        let more_sizes_raw = try!(input_stream.read_exact((4 * (segment_count & !1)) as uint));
-        for ii in range(0, segment_count as uint - 1) {
+        let more_sizes_raw = try!(input_stream.read_exact((4 * (segment_count & !1)) as usize));
+        for ii in range(0, segment_count as usize - 1) {
             let size = unsafe {
                 let p : *const WireValue<u32> =
                     std::mem::transmute(more_sizes_raw.get_unchecked(ii * 4));
@@ -89,8 +89,8 @@ pub fn new_reader<U : std::io::Reader>(input_stream : &mut U,
                               receiving end, see capnp::ReaderOptions.");
     }
 
-    let mut owned_space : Vec<Word> = allocate_zeroed_words(total_words as uint);
-    let buf_len = total_words as uint * BYTES_PER_WORD;
+    let mut owned_space : Vec<Word> = allocate_zeroed_words(total_words as usize);
+    let buf_len = total_words as usize * BYTES_PER_WORD;
 
     unsafe {
         let ptr : *mut u8 = std::mem::transmute(owned_space.as_mut_slice().as_mut_ptr());
@@ -100,20 +100,20 @@ pub fn new_reader<U : std::io::Reader>(input_stream : &mut U,
 
     // TODO(maybe someday) lazy reading like in capnp-c++?
 
-    let mut segment_slices : Vec<(uint, uint)> = vec!((0, segment0_size as uint));
+    let mut segment_slices : Vec<(usize, usize)> = vec!((0, segment0_size as usize));
 
     let arena = {
-        let segment0 : &[Word] = owned_space.slice(0, segment0_size as uint);
+        let segment0 : &[Word] = owned_space.slice(0, segment0_size as usize);
         let mut segments : Vec<&[Word]> = vec!(segment0);
 
         if segment_count > 1 {
             let mut offset = segment0_size;
 
-            for ii in range(0, segment_count as uint - 1) {
-                segments.push(owned_space.slice(offset as uint,
-                                               (offset + more_sizes.as_slice()[ii]) as uint));
-                segment_slices.push((offset as uint,
-                                     (offset + more_sizes.as_slice()[ii]) as uint));
+            for ii in range(0, segment_count as usize - 1) {
+                segments.push(owned_space.slice(offset as usize,
+                                               (offset + more_sizes.as_slice()[ii]) as usize));
+                segment_slices.push((offset as usize,
+                                     (offset + more_sizes.as_slice()[ii]) as usize));
                 offset += more_sizes.as_slice()[ii];
             }
         }
@@ -136,12 +136,12 @@ pub fn write_message<T : std::io::Writer, U : MessageBuilder>(
     try!(message.get_segments_for_output(
         |segments| {
 
-            let table_size : uint = (segments.len() + 2) & (!1);
+            let table_size : usize = (segments.len() + 2) & (!1);
 
             let mut table : Vec<WireValue<u32>> = Vec::with_capacity(table_size);
             unsafe { table.set_len(table_size) }
 
-            table.as_mut_slice()[0u].set((segments.len() - 1) as u32);
+            table.as_mut_slice()[0].set((segments.len() - 1) as u32);
 
             for i in range(0, segments.len()) {
                 table.as_mut_slice()[i + 1].set(segments[i].len() as u32);
